# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RESQHER Women Safety Emergency API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SendAlertBody {
  userId: string;
  locationLink: string;
  timestamp: string;
  recipientPhone: string;
  /** @nullable */
  userName?: string | null;
}

export interface SendAlertResponse {
  success: boolean;
  firebaseKey?: string;
  /** @nullable */
  smsSid?: string | null;
  message: string;
}

export interface EmergencyRecord {
  id: string;
  userId: string;
  timestamp: string;
  locationLink: string;
  status: string;
  /** @nullable */
  localVideoPath?: string | null;
}

export interface EmergencyListResponse {
  emergencies: EmergencyRecord[];
}

export interface ErrorResponse {
  error: string;
}