# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RESQHER Women Safety Emergency API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Store emergency metadata in Firebase and send SMS via Twilio
 * @summary Send emergency alert
 */
export const SendAlertBody = zod.object({
  userId: zod.string(),
  locationLink: zod.string(),
  timestamp: zod.string(),
  recipientPhone: zod.string(),
  userName: zod.string().nullish(),
});

export const SendAlertResponse = zod.object({
  success: zod.boolean(),
  firebaseKey: zod.string().optional(),
  smsSid: zod.string().nullish(),
  message: zod.string(),
});

/**
 * Get all emergency alerts for a specific user
 * @summary Get emergency history
 */
export const GetEmergenciesParams = zod.object({
  userId: zod.coerce.string(),
});

export const GetEmergenciesResponse = zod.object({
  emergencies: zod.array(
    zod.object({
      id: zod.string(),
      userId: zod.string(),
      timestamp: zod.string(),
      locationLink: zod.string(),
      status: zod.string(),
      localVideoPath: zod.string().nullish(),
    }),
  ),
});